CI: Split apt dependency lists and caches
Splitting the dependency lists is primarily to keep closer track of
which dependencies are needed for which jobs.

For this to work properly with caching the packages in gitlab, we also
give each job its own cache. This potentially means a little more
storage in our cache object store due to overlapping caches, though also
means smaller cache up and downloads for individual jobs.