Make it harder to create broken TorEnviron hierarchy
chutney currently depends on changes in the networks `_dfltEnv` to be
reflected in the environments of nodes that have already been created.
i.e. `_dfltEnv` should be an ancestor environment to all node
environments.

Taking the network in the `Node` initializer makes this difficult to get
wrong.

Since the Network must now be created before creating any Nodes, we
also:
* Expose `Network._addNode` as `Network.AddNode`
* Add a convenience method `Network.AddNodes`
* Remove `createNetwork`