Traffic.AsynChatProducer.more: don't allow None
The documentation in the asynchat module for the interface that this
represents doesn't permit returning None; end of the stream is indicated
with an empty byte stream.

While the asynchat appears to treat None as synonymous with the empty
byte sequence in practice, it's hard to be sure that it does so for all
versions.