torrc_templates: specifically take a TorEnviron
Both the previous `Environ` type annotation and the new `TorEnviron`
aren't actually quite correct, since the templating wraps the
`TorEnviron` in some other helpers. It's not a problem at runtime
though, and mypy accepts it since it can't see through some of the
metaprogramming.

This a step towards migrating `TorEnviron` to be a dataclass, and the
"template" functions to not actually invoke the templating code.