Have Node take a NodeConfig instead of creating one
This means that network config scripts now create NodeConfigs instead of
Nodes, and that the actual Nodes are created only later when actually
adding to the network.

This means that the normal path for creating NodeConfig can now be
type-checked, since the arguments are no longer being funneled through a
`**kwargs: Any` in Node's `__init__`.

This is also part of clearing the way to move fields that aren't actually
intended as inputs (e.g. `id`) out of `NodeConfig`.