Fix bridges-obfs4 network and multi-phase configurations
It was using `Environment.set_runtime` to dynamically teach chutney
to build obfs4 pt bridge lines. `Environment` is gone altogether,
so updated by building the obfs4 knowledge into chutney itself.

This also fixes a bug in which `network.dir` wasn't getting initialized
in cases where phase 1 of configuration isn't executed. (Running a
subsequent phase of configuration tries to use it, and was crashing).