Begin merging torrc templates into common.i, starting with client.tmpl
At a high level, we want all of the interesting specification of how a
node should be configured to go into NodeConfig rather than in torrc
templates, so that NodeConfig can be used to generate arti configs as
well.

To begin, we move all instances of setting `SocksPort` into `common.i`.

* modify `Node.socksport` to return an `Option[int]`,
  returning None if the `NodeConfig.client` is False.
* Add `SocksPort` directive into `common_i.py`'s output,
  setting to 0 if `Node.socksport.is_none()`.
* In other templates that set `SocksPort`, remove it, and
  assert that `SocksPort` is still set as expected in the output.