Merge client-only-v6-md.i into common.i, and explicitly unset ip
Previously disabling ipv4 was expressed only through the template
selection.

Now it's expressed as `ip=None`, and we raise an exception if one of the
templates that previously meant "ipv6 only" doesn't has `ip=None`.

We avoid breaking network scripts by setting `ip=None` in `NodeWrapper`
for torrc templates that now require it.