Create and use BridgeDescriptors
Instead of pre-rendering Bridge lines from bridges, create higher level
BridgeDescriptors and push rendering the text Bridge line with the other
config generation code.

We also drop `Network._checkConfig`, which was checking the generated
configs before finding the Bridge lines. I'm not actually sure how this
was passing before - tor rejects a config with UseBridges but without
and Bridge lines. Anyway, this validation was redundant with validation
we do shortly later when we actually generate the real torrc's after
gathering the bridge and authority info.