Update bwscanner network to not spuriously set exit=1
This makes it more consistent with the version of this network
that sbws actually uses, where 'exit=1' is not set:
<https://gitlab.torproject.org/tpo/network-health/sbws/-/blob/c2ead58ec8dc4f0eaa7c9f530393d4766e839d09/tests/integration/chutney_data/bwscanner>.
The reference issue #40013 also makes the intent a bit clearer that this
is trying to configure these relays to *not* be exits.

Now that chutney uses the `exit` option to decide whether to configure
a relay as an exit, leaving in the `exit=1` actually explicitly *does*
configure these to be exits, which is an undesired change.

Note though that as far as I know this network is only ~used by the sbws
CI, which actually overwrites it with its own modified version anyway.

That CI appears to currently be broken, in part due to also copying in
custom templates which already no longer works since templates are now
python modules instead of data files. That will be easier to fix after
merging this MR, in which case the desired customizations can be added
directly to the network file via the `extra_raw_torrc` option.