Add and run tests/network-config-tests-diff
This test runs ./tests/network-config-tests (which attempts to configure
every built-in chutney network), first at HEAD, then at a specified
commit (using a cloned repo in a tmpdir). Both runs are executed under
shadow to get determinism. Finally the outputs are compared using
diffoscope.

In CI the test fails if there is any diff, but the job is marked as
`allow_failure: true`. The intent is to raise awareness when there is a
diff, and show what it is, but not to block merging.

The test can also be run locally, assuming diffoscope and shadow are
installed.

Fixes #40027