networks/bwscanner: don't override nickname
Doing so confuses other chutney logic. e.g. when waiting for bootstrap,
chutney doesn't know to look for the overridden name instead of the one
it generated.

We have to also shorten the corresponding tags, which are used to
generate the name, since otherwise tor rejects the long generated name.