Move isRunning to NodeController and don't take pid
We call this method via the base class, so it belongs there, but it may
not make sense to take a pid for some implementations of NodeController
(e.g. a "RemoteNodeController").

It is still useful to be able to provide the pid internally in
LocalNodeController, so we move most of the internal implementation
details to a new function LocalNodeController._is_running_with_pid.