Add DirFormat enum and replace corresponding strings
This is part of trying to understand and better-structure the various
aggregate summaries that are processed in the controller.

This also extracts some list filtering functionality from
`combineDirInfoStatuses` into a separate utility function
`values_for_keys`. This is in this commit in particular to make the type
signatures work out more nicely, and generally seems a reasonable way to
extract some orthogonal functionality.