Replace NodeController.getNodeCacheDirInfoPaths with check_node_in_dirinfo
The design of `getNodeCacheDirInfoPaths` pre-supposes that directory
data are all stored as individual files on disk. This is not the case
for arti, where e.g. descriptors and microdescriptors are stored
directly in a sqlite database.

We drop `getNodeCacheDirInfoPaths` from the NodeController base class,
and replace it with `check_node_in_dirinfo`, which provides the
functionality that `getNodeCacheDirInfoPaths` was ultimately used for -
to check whether some other node is present in the current node's
directory data.

A couple-other changes that got munged in to support this change:

* Add `NodeController.getEd25519Id`. To implement
  `check_node_in_dirinfo` we need to be able to get the other node's
  ed25519 id.
* `NodeController.getNodeDirInfoStatus` - drop the string description
  from the return type. Otherwise we would have needed to return it from
  `check_node_in_dirinfo`, but it's redundant with the
  `DirInfoStatusCode` it already returns.
* Move `LocalNodeController.getNodeDirInfoStatusPattern` to
  `DirFormat.status_pattern`, since it's a property of the directory
  format and not of the controller.